+++ Locadora/locadora.hpp
@@ -63,7 +63,7 @@ public:
   void aluguel(long long cpf, vector<int> id);
   void devolucao(long long cpf, int dias);
   void recomendarFilmes(long long cpf);
-  void avaliarFilme(int id, int nota);
+  void avaliarFilme(int id, float nota);
   vector<Filme*> filmes;
   vector<Cliente*> clientes;
   vector<Cliente*> clientesOrdenados;
